resources/ApplicationAutoscaling_ScheduledAction/requirements.txt
attrs==22.1.0 ; python_version >= "3.7" and python_version < "4.0"
aws-encryption-sdk==3.1.0 ; python_version >= "3.7" and python_version < "4.0"
boto3==1.24.92 ; python_version >= "3.7" and python_version < "4.0"
botocore==1.27.92 ; python_version >= "3.7" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.7" and python_version < "4.0"
cloudformation-cli-python-lib==2.1.12 ; python_version >= "3.7" and python_version < "4.0"
cryptography==38.0.1 ; python_version >= "3.7" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.7" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.7" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.7" and python_version < "4.0"
s3transfer==0.6.0 ; python_version >= "3.7" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.7" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.7" and python_version < "4"
wrapt==1.14.1 ; python_version >= "3.7" and python_version < "4.0"
